@@ -11,13 +11,16 @@ platforms:
   web:
   
 environment:
-  sdk: ">=2.14.0 <3.0.0"
+  sdk: ">=2.19.2 <3.0.0"
   flutter: ">=2.5.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  path: ^1.8.0
+  path: ^1.8.2
+  universal_io: ^2.2.0
+  webview_flutter: ^4.0.6
+  webview_flutter_web: ^0.2.2+1
 
 dev_dependencies:
   flutter_test:
